Refactor: Move significantly complex reducers to separate modules
